separated schema file in own directory
